// src/app/models/enums/etat-compet.enum.ts
export enum EtatCompet {
  START_QUESTION_NORMALE = 'START_QUESTION_NORMALE',
  QUESTION_NORMALE_POSEE = 'QUESTION_NORMALE_POSEE',
  VOTES_OUVERTS = 'VOTES_OUVERTS',
  VOTES_FERMES = 'VOTES_FERMES',
  BONNE_REPONSE_AFFICHEE = 'BONNE_REPONSE_AFFICHEE',
  START_QUESTION_SUPER_CASH = 'START_QUESTION_SUPER_CASH',
  QUESTION_SUPER_CASH_CHOISIE = 'QUESTION_SUPER_CASH_CHOISIE',
  QUESTION_SUPER_CASH_POSEE = 'QUESTION_SUPER_CASH_POSEE',
  CHRONO_SUPER_CASH_LANCE = 'CHRONO_SUPER_CASH_LANCE',
  REPONSE_SUPER_CASH_DONNEE = 'REPONSE_SUPER_CASH_DONNEE',
  BONNE_REPONSE_SUPER_CASH_AFFICHEE = 'BONNE_REPONSE_SUPER_CASH_AFFICHEE',
}